# openapi.yaml
openapi: 3.0.3
info:
  title: KarenFlix API
  version: "1.0.0"
  description: |
    API REST para registrar, calificar y rankear películas, animes y series geek.
    Autenticación con JWT (Bearer), validaciones con express-validator, rate limiting,
    transacciones reales en MongoDB y documentación Swagger.
servers:
  - url: http://localhost:4000
    description: Servidor local (desarrollo)
  - url: https://{host}
    description: Servidor desplegado
    variables:
      host:
        default: api.karenflix.example.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    FieldError:
      type: object
      properties:
        field: { type: string, example: "title" }
        msg: { type: string, example: "Title is required" }
    ErrorResponse:
      type: object
      properties:
        ok: { type: boolean, example: false }
        message: { type: string, example: "Validation error" }
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
...
    UserPublic:
      type: object
      properties:
        id: { type: string, example: "6639c8dc9f7a2d5c2f4d1a00" }
        email: { type: string, format: email }
        displayName: { type: string, example: "Karen" }
        role:
          type: string
          enum: [user, admin]
        createdAt: { type: string, format: date-time }
    AuthRegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        displayName: { type: string }
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    AuthTokenResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        tokenType: { type: string, example: "Bearer" }
        accessToken: { type: string, example: "eyJhbGciOiJIUzI1NiIsInR5cCI..." }
        expiresIn: { type: integer, example: 86400 }
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Registro de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthRegisterRequest" }
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  user: { $ref: "#/components/schemas/UserPublic" }
                  accessToken:
                    type: string
                    description: "Opcional: si el backend inicia sesión tras registrar"
        "400":
          description: Error de validación
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Inicio de sesión
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthLoginRequest" }
      responses:
        "200":
          description: Token emitido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokenResponse" }
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...